64bit fixes for vairant builder.
(bzr r3008.2.49)

+++ UnityCore/Variant.h
@@ -49,10 +49,10 @@ public:
   ~Variant();
 
   std::string GetString() const;
-  gint32 GetInt32() const;
-  guint32 GetUInt32() const;
-  gint64 GetInt64() const;
-  guint64 GetUInt64() const;
+  int32_t GetInt32() const;
+  uint32_t GetUInt32() const;
+  int64_t GetInt64() const;
+  uint64_t GetUInt64() const;
   bool GetBool() const;
   double GetDouble() const;
   float GetFloat() const;
@@ -83,10 +83,10 @@ public:
   BuilderWrapper& add(char const* name, bool value);
   BuilderWrapper& add(char const* name, char const* value);
   BuilderWrapper& add(char const* name, std::string const& value);
-  BuilderWrapper& add(char const* name, gint32 value);
-  BuilderWrapper& add(char const* name, gint64 value);
-  BuilderWrapper& add(char const* name, guint32 value);
-  BuilderWrapper& add(char const* name, guint64 value);
+  BuilderWrapper& add(char const* name, int32_t value);
+  BuilderWrapper& add(char const* name, uint32_t value);
+  BuilderWrapper& add(char const* name, int64_t value);
+  BuilderWrapper& add(char const* name, uint64_t value);
   BuilderWrapper& add(char const* name, float value);
   BuilderWrapper& add(char const* name, double value);
   BuilderWrapper& add(char const* name, GVariant* value);
